Add a PMC ID converter query
When a pmid is missing but doi is present, we will ask PMC id
converter about the doi. If it returns a matching pmid, we will use
that with the pubmed's literature citation exporter.

This introduces a configuration file (currently relative path) to
store user's e-mail used for (polite) queries to the API. Uses tomllib
(built-in since 3.11) to read the configuration file.