Add a bibliographic crossref query
This adds the ability to perform a bibliographic query in crossref
api, i.e. query using a citation text. This is useful when there are
no identifiers available.

My previous experiences with parsing sfb publications suggest that
evaluating first few results may be necessary to discard commentaries
or preprints that share e.g. the same title.

Returns the same kind of metadata as a works query by doi, meaning we
need to unpack some single-item lists to get compatibility with csl.