Add a function for finding id in url
The most recent version of a publication list that I got had hardly
any identifiers given directly; instead it contained urls in various
forms (journal websites, journal pdfs, pubmed, pubmed
central). Thankfully, about half of those contain DOIs or
Pubmed (Central) IDs in some shape of form that can be extracted with
regex.

This adds a function for finding an id in a url. Although it's all
regex, it felt more natural to me to keep it separate from a function
that finds an id in its "canonical" form, i.e. e.g. "doi: ...", that
can be used to parse a free-form citation.