Simplify internal representation, standardise on csl
After doing a bibliographic query with crossref and finding the best
result, we take its doi and perform an additional query to doi.org
with content negotiation to get a csl+json response.

This is slightly inefficient, because the requests will most likely be
redirected back to crossref. However, the REST api that we use for
bibliographic queries, seems to only return crossref-api-message+json
format. I may be wrong, but I did not dig deeper into crossref
api(s). Although the formats are very close to each other, making
another query is simpler than translation.

We remove the Publication class, and instead use regular dictionaries
returned by the queries, to which we add an additional field for
comments.